CLN: move pandas.tseries.period -> pandas.core.indexes.period
