+++ pandas/core/api.py
@@ -29,7 +29,7 @@ from pandas.tseries.tools import to_datetime
 from pandas.tseries.index import (DatetimeIndex, Timestamp,
                                   date_range, bdate_range)
 from pandas.tseries.tdi import TimedeltaIndex, Timedelta
-from pandas.tseries.period import Period, PeriodIndex
+from pandas.core.indexes.period import Period, PeriodIndex
 from pandas.core.resample import TimeGrouper
 
 # see gh-14094.
