+++ pandas/plotting/_timeseries.py
@@ -3,11 +3,11 @@
 import numpy as np
 
 from matplotlib import pylab
-from pandas.tseries.period import Period
+from pandas.core.indexes.period import Period
 from pandas.tseries.offsets import DateOffset
 import pandas.tseries.frequencies as frequencies
 from pandas.tseries.index import DatetimeIndex
-from pandas.tseries.period import PeriodIndex
+from pandas.core.indexes.period import PeriodIndex
 from pandas.tseries.tdi import TimedeltaIndex
 from pandas.io.formats.printing import pprint_thing
 import pandas.compat as compat
