+++ pandas/tests/frame/test_timeseries.py
@@ -547,7 +547,7 @@ class TestDataFrameTimeSeriesMethods(tm.TestCase, TestData):
 
     def test_frame_to_period(self):
         K = 5
-        from pandas.tseries.period import period_range
+        from pandas.core.indexes.period import period_range
 
         dr = date_range('1/1/2000', '1/1/2001')
         pr = period_range('1/1/2000', '1/1/2001')
