+++ pandas/tests/indexes/period/test_ops.py
@@ -4,7 +4,7 @@ from datetime import timedelta
 import pandas as pd
 import pandas._libs.tslib as tslib
 import pandas.util.testing as tm
-import pandas.tseries.period as period
+import pandas.core.indexes.period as period
 from pandas import (DatetimeIndex, PeriodIndex, period_range, Series, Period,
                     _np_version_under1p10, Index, Timedelta, offsets)
