+++ pandas/tests/indexes/period/test_setops.py
@@ -2,7 +2,7 @@ import numpy as np
 
 import pandas as pd
 import pandas.util.testing as tm
-import pandas.tseries.period as period
+import pandas.core.indexes.period as period
 from pandas import period_range, PeriodIndex, Index, date_range
