+++ pandas/tests/test_resample.py
@@ -21,7 +21,7 @@ from pandas.tseries.frequencies import MONTHS, DAYS
 from pandas.tseries.frequencies import to_offset
 from pandas.tseries.index import date_range
 from pandas.tseries.offsets import Minute, BDay
-from pandas.tseries.period import period_range, PeriodIndex, Period
+from pandas.core.indexes.period import period_range, PeriodIndex, Period
 from pandas.core.resample import (DatetimeIndex, TimeGrouper,
                                   DatetimeIndexResampler)
 from pandas.tseries.tdi import timedelta_range, TimedeltaIndex
