+++ pandas/tseries/common.py
@@ -14,7 +14,7 @@ from pandas.core.dtypes.common import (
 from pandas.core.base import PandasDelegate, NoNewAttributesMixin
 from pandas.tseries.index import DatetimeIndex
 from pandas._libs.period import IncompatibleFrequency  # noqa
-from pandas.tseries.period import PeriodIndex
+from pandas.core.indexes.period import PeriodIndex
 from pandas.tseries.tdi import TimedeltaIndex
 from pandas.core.algorithms import take_1d
